Updated the folder path for Set-SecurityPolicy
The security policies were not deployed. This was caused by processing the folder path twice causing the folder name to be added to the path twice. I.e., /foldername becomes /foldername/foldername.

Each publish function already processes the path into the $currenFolder variable. There is no need to process it a second time in the Set-SecurityPolicy function.